+++ 01-node-tutorial/answers/practice2.js
@@ -1,0 +1,7 @@
+// Shorter and cleaner approach:
+module.exports.sentence = 'I am learning NodeJS pretty fast using:';
+
+// A longer approach:
+// const sentence = 'I am learning NodeJS pretty fast!';
+
+// module.exports = sentence;
